AbstractRepository with support for creating and updating LiveDatas
