+++ app/src/main/java/de/theess/eisbaer/data/NoteRepository.kt
@@ -1,55 +1,16 @@
 package de.theess.eisbaer.data
 
 import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import de.theess.eisbaer.EisbaerApplication
 import io.requery.kotlin.desc
 import io.requery.kotlin.eq
 import io.requery.kotlin.like
 import timber.log.Timber
-import java.util.*
 
-typealias NoteListProviderFunction = () -> List<Note>?
-
-class NoteRepository private constructor(private val database: Database) : DatabaseListener {
-
-    /**
-     * Map of LiveData to value function. When the database changes, the livedatas will be updated
-     * using the value function.
-     */
-    private val liveDataMap: MutableMap<MutableLiveData<List<Note>>, NoteListProviderFunction> =
-        WeakHashMap<MutableLiveData<List<Note>>, NoteListProviderFunction>()
-
-    init {
-        database.addListener(this)
-    }
-
-    /**
-     * Stores the function and a new LiveData in the liveDataMap, and updates the LiveData using
-     * the function result.
-     */
-    private fun wrap(valueProvider: NoteListProviderFunction): LiveData<List<Note>> {
-        val liveData = MutableLiveData<List<Note>>()
-        liveDataMap[liveData] = valueProvider
-
-        valueProvider()?.let {
-            liveData.postValue(it)
-        }
-
-        return liveData
-    }
-
-    override fun update() {
-        Timber.d("Updating %d listeners.", liveDataMap.size)
-        // Update each LiveData using the provider function.
-        liveDataMap.forEach {
-            Timber.d("Updating note listener. Observers: %b", it.key.hasObservers())
-            it.key.postValue(it.value())
-        }
-    }
+class NoteRepository(private val database: Database) : AbstractRepository(database) {
 
     fun getAll(): LiveData<List<Note>> {
-        return wrap {
+        return toLiveData {
             database.store?.run {
                 select(Note::class)
                     .where(Note::trashed.eq(0))
@@ -60,7 +21,7 @@ class NoteRepository private constructor(private val database: Database) : Datab
 
     fun query(query: String): LiveData<List<Note>> {
         Timber.d("query: $query")
-        return wrap {
+        return toLiveData {
             database.store
                 ?.run {
                     select(Note::class)
@@ -74,10 +35,10 @@ class NoteRepository private constructor(private val database: Database) : Datab
 
     fun getById(id: Long): LiveData<Note> {
         Timber.d("getById: $id")
-        return database.store
-            ?.findByKey(Note::class, id)
-            ?.let { MutableLiveData(it) }
-            ?: MutableLiveData()
+        return toLiveData {
+            database.store
+                ?.findByKey(Note::class, id)
+        }
     }
 
     companion object {
